+++ framework/db/Command.php
@@ -1030,7 +1030,7 @@ class Command extends Component
 
         return $this->setSql($sql)->requireTableSchemaRefresh($viewName);
     }
-    
+
     /**
      * Drops a SQL View.
      *
@@ -1128,7 +1128,7 @@ class Command extends Component
                 ];
                 $result = $cache->get($cacheKey);
                 if (is_array($result) && isset($result[0])) {
-                    Yii::trace('Query result served from cache', 'yii\db\Command::query');
+                    Yii::debug('Query result served from cache', 'yii\db\Command::query');
                     return $result[0];
                 }
             }
@@ -1159,7 +1159,7 @@ class Command extends Component
 
         if (isset($cache, $cacheKey, $info)) {
             $cache->set($cacheKey, [$result], $info[1], $info[2]);
-            Yii::trace('Saved query result in cache', 'yii\db\Command::query');
+            Yii::debug('Saved query result in cache', 'yii\db\Command::query');
         }
 
         return $result;