+++ framework/db/Query.php
@@ -9,8 +9,8 @@ namespace yii\db;
 
 use Yii;
 use yii\base\Component;
+use yii\base\InvalidArgumentException;
 use yii\base\InvalidConfigException;
-use yii\base\InvalidParamException;
 
 /**
  * Query represents a SELECT SQL statement in a way that is independent of DBMS.
@@ -554,7 +554,7 @@ PATTERN;
 
             if ($tableName instanceof Expression) {
                 if (!is_string($alias)) {
-                    throw new InvalidParamException('To use Expression in from() method, pass it in array format with alias.');
+                    throw new InvalidArgumentException('To use Expression in from() method, pass it in array format with alias.');
                 }
                 $cleanedUpTableNames[$this->ensureNameQuoted($alias)] = $tableName;
             } elseif ($tableName instanceof self) {