+++ framework/db/mysql/QueryBuilder.php
@@ -7,7 +7,7 @@
 
 namespace yii\db\mysql;
 
-use yii\base\InvalidParamException;
+use yii\base\InvalidArgumentException;
 use yii\base\NotSupportedException;
 use yii\db\Exception;
 use yii\db\Expression;
@@ -165,7 +165,7 @@ class QueryBuilder extends \yii\db\QueryBuilder
      * @param mixed $value the value for the primary key of the next new row inserted. If this is not set,
      * the next new row's primary key will have a value 1.
      * @return string the SQL statement for resetting sequence
-     * @throws InvalidParamException if the table does not exist or there is no sequence associated with the table.
+     * @throws InvalidArgumentException if the table does not exist or there is no sequence associated with the table.
      */
     public function resetSequence($tableName, $value = null)
     {
@@ -181,10 +181,10 @@ class QueryBuilder extends \yii\db\QueryBuilder
 
             return "ALTER TABLE $tableName AUTO_INCREMENT=$value";
         } elseif ($table === null) {
-            throw new InvalidParamException("Table not found: $tableName");
+            throw new InvalidArgumentException("Table not found: $tableName");
         }
 
-        throw new InvalidParamException("There is no sequence associated with table '$tableName'.");
+        throw new InvalidArgumentException("There is no sequence associated with table '$tableName'.");
     }
 
     /**